feat: Refactor HomeViewModel and related components for improved state management and liking functionality
- Created new `HomeViewModel`, `HomeViewStates`, and `HomeEvent` in the `screens.home` package.
- Removed the old `HomeViewModel`, `HomeViewStates`, and `HomeEvent` from the root `presentation` package.
- Implemented `reloadLikedExperiences` in `HomeViewModel` to reload liked experiences and update the state of displayed experiences.
- Updated `HomeViewModel` to manage the loading state and handle errors.
- Added new methods in `HomeViewModel` to handle experience resources and update experiences with their liked status.
- Implemented a debouncing mechanism for the like button in `HomeViewModel`.
- Added logic to update recommended, most recent, and search result lists when an experience is liked.
- Updated `HomeScreen` to trigger `reloadLikedExperiences` on launch.
